% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WHOCountryNames.R
\name{WHOCountryNames}
\alias{WHOCountryNames}
\title{WHO_Names function
This function will transform any country name as per WHO convention}
\usage{
WHOCountryNames(dataset, countryfield)
}
\arguments{
\item{dataset}{\itemize{
\item dataset containing country names to convert to WHO reference
}}

\item{countryfield}{\itemize{
\item name of the field containing country names - Don't use quotes!
}}
}
\description{
WHO_Names function
This function will transform any country name as per WHO convention
}
